refactor: extract icon path and app discovery helpers from main
Made-with: Cursor
